refactor: make creating builder infallible
