+++ rust/src/operations/transaction.rs
@@ -148,8 +148,7 @@ async fn do_transaction(
     app_metadata: Option<serde_json::Map<String, serde_json::Value>>,
     context: Arc<TaskContext>,
 ) -> DataFusionResult<SendableRecordBatchStream> {
-    let mut table = DeltaTableBuilder::try_from_uri(table_uri)
-        .map_err(to_datafusion_err)?
+    let mut table = DeltaTableBuilder::from_uri(table_uri)
         .build()
         .map_err(to_datafusion_err)?;
     let schema = input.schema().clone();
