+++ rust/src/operations/write.rs
@@ -177,8 +177,7 @@ async fn do_write(
     mode: SaveMode,
     context: Arc<TaskContext>,
 ) -> DataFusionResult<SendableRecordBatchStream> {
-    let mut table = DeltaTableBuilder::try_from_uri(&table_uri)
-        .map_err(to_datafusion_err)?
+    let mut table = DeltaTableBuilder::from_uri(&table_uri)
         .build()
         .map_err(to_datafusion_err)?;
     let metrics = ExecutionPlanMetricsSet::new();
