+++ rust/src/writer/record_batch.rs
@@ -76,7 +76,7 @@ impl RecordBatchWriter {
         partition_columns: Option<Vec<String>>,
         storage_options: Option<HashMap<String, String>>,
     ) -> Result<Self, DeltaWriterError> {
-        let storage = DeltaTableBuilder::try_from_uri(&table_uri)?
+        let storage = DeltaTableBuilder::from_uri(&table_uri)
             .with_storage_options(storage_options.unwrap_or_default())
             .build_storage()?;
